@@ -1,34 +1,36 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-    <UserSecretsId>aspnet-OpenSourceSCORMLMS-216CDD04-809C-4E90-8838-B15D3F9DACB5</UserSecretsId>
-  </PropertyGroup>
+    <PropertyGroup>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <UserSecretsId>aspnet-OpenSourceSCORMLMS-216CDD04-809C-4E90-8838-B15D3F9DACB5</UserSecretsId>
+        <LangVersion>8</LangVersion>
+        <AspNetCoreModuleName>AspNetCoreModuleV2</AspNetCoreModuleName>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <Content Remove="appsettings.Production..json" />
-  </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="Bower" Version="1.3.11"/>
+        <PackageReference Include="Karambolo.Extensions.Logging.File" Version="3.1.1"/>
+        <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="3.1.4"/>
+        <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="3.1.4"/>
+        <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="3.1.4"/>
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.4"/>
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="3.1.4"/>
+        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.4"/>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.4"/>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="3.1.4"/>
+        <PackageReference Include="Newtonsoft.Json" Version="12.0.3"/>
+        <PackageReference Include="System.IO.Compression.ZipFile" Version="4.3.0"/>
+    </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Bower" Version="1.3.11" />
-    <PackageReference Include="Karambolo.Extensions.Logging.File" Version="2.1.1" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.0" PrivateAssets="All" />
-    <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
-    <PackageReference Include="System.IO.Compression.ZipFile" Version="4.3.0" />
-  </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\SCORM_XMLObjects\V1_2\adlcp_rootv1p2\adlcp_rootv1p2.csproj"/>
+        <ProjectReference Include="..\SCORM_XMLObjects\V1_2\Altova\Altova.csproj"/>
+        <ProjectReference Include="..\SCORM_XMLObjects\V1_2\imscp_rootv1p1p2\imscp_rootv1p1p2.csproj"/>
+        <ProjectReference Include="..\SCORM_XMLObjects\V1_2\imsmd_rootv1p2p1\imsmd_rootv1p2p1.csproj"/>
+        <ProjectReference Include="..\SCORM_XMLObjects\V1_3\adlcp_v1p3\adlcp_v1P3.csproj"/>
+    </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Areas\Identity\Services\" />
-    <Folder Include="Data\Migrations\" />
-    <Folder Include="SCORMPackages\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SCORM_XMLObjects\V1_2\adlcp_rootv1p2\adlcp_rootv1p2.csproj" />
-    <ProjectReference Include="..\SCORM_XMLObjects\V1_2\Altova\Altova.csproj" />
-    <ProjectReference Include="..\SCORM_XMLObjects\V1_2\imscp_rootv1p1p2\imscp_rootv1p1p2.csproj" />
-    <ProjectReference Include="..\SCORM_XMLObjects\V1_2\imsmd_rootv1p2p1\imsmd_rootv1p2p1.csproj" />
-    <ProjectReference Include="..\SCORM_XMLObjects\V1_3\adlcp_v1p3\adlcp_v1P3.csproj" />
-  </ItemGroup>
+    <ItemGroup>
+        <Folder Include="wwwroot\SCORMCourses"/>
+    </ItemGroup>
 </Project>
