@@ -13,10 +13,10 @@ namespace OpenSourceSCORMLMS.Pages
         public string UserID { get; set; }
         private readonly UserManager<IdentityUser> _userManager;
         private IConfiguration _configuration;
-        private IHostingEnvironment _environment;
+        private IWebHostEnvironment _environment;
         private ILogger _logger { get; set; }
         private Helpers.DatabaseHelper databaseHelper { get; set; }
-        public MyCoursesForStudyModel(UserManager<IdentityUser> User, IConfiguration Configuration, IHostingEnvironment hostingEnvironment, ILogger<UploadFileModel> logger)
+        public MyCoursesForStudyModel(UserManager<IdentityUser> User, IConfiguration Configuration, IWebHostEnvironment hostingEnvironment, ILogger<UploadFileModel> logger)
         {
             _userManager = User;
             _configuration = Configuration;
