@@ -16,11 +16,12 @@ namespace OpenSourceSCORMLMS.Pages
 
         private readonly UserManager<IdentityUser> _userManager;
         private IConfiguration _configuration;
-        private IHostingEnvironment _environment;
+        private IWebHostEnvironment _environment;
         private ILogger _logger { get; set; }
         private Helpers.DatabaseHelper databaseHelper { get; set; }
         private int SCORM_Course_id;
-        public PackageModel(UserManager<IdentityUser> User, IConfiguration Configuration, IHostingEnvironment hostingEnvironment, ILogger<UploadFileModel> logger)
+
+        public PackageModel(UserManager<IdentityUser> User, IConfiguration Configuration, IWebHostEnvironment hostingEnvironment, ILogger<UploadFileModel> logger)
         {
             _userManager = User;
             _configuration = Configuration;
@@ -28,13 +29,15 @@ namespace OpenSourceSCORMLMS.Pages
             _logger = logger;
             databaseHelper = new Helpers.DatabaseHelper(_logger);
         }
-        [Authorize]
+
+        // [Authorize]
         public void OnGet()
         {
             if (!Models.SignedInUser.isSignedIn)
             {
                 Response.Redirect("/Identity/Account/Login?returnUrl=" + Request.Path);
             }
+
             string UserID = _userManager.GetUserId(HttpContext.User);
             if (Request.Query["id"] != String.Empty && Helpers.UtilityFunctions.isInteger(Request.Query["id"]))
             {
@@ -43,18 +46,20 @@ namespace OpenSourceSCORMLMS.Pages
                 bUserHasCourse = databaseHelper.isCourseInUserStudyArea(SCORM_Course_id, UserID);
             }
         }
-        [Authorize]
+
+        // [Authorize]
         public IActionResult OnPost([FromForm] Models.SCORMCourse package)
         {
             string UserID = _userManager.GetUserId(HttpContext.User);
             // determine if course is already in their study area
             // if so, just go to study area
             // if not, add to study area
-            if (! databaseHelper.isCourseInUserStudyArea(package.id, UserID))
+            if (!databaseHelper.isCourseInUserStudyArea(package.id, UserID))
             {
                 // add course to their study area
                 databaseHelper.AddCourseToUsersStudyArea(package.id, UserID);
             }
+
             return RedirectToPage("MyCoursesForStudy");
         }
     }
