@@ -1,3 +1,4 @@
+using System;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
@@ -10,6 +11,8 @@ using Microsoft.Extensions.FileProviders;
 using Newtonsoft.Json.Serialization;
 using OpenSourceSCORMLMS.Data;
 using System.IO;
+using Microsoft.AspNetCore.Http.Features;
+using Microsoft.Extensions.Hosting;
 
 namespace OpenSourceSCORMLMS
 {
@@ -31,55 +34,98 @@ namespace OpenSourceSCORMLMS
                 options.CheckConsentNeeded = context => true;
                 options.MinimumSameSitePolicy = SameSiteMode.None;
             });
-            services.AddSingleton<IConfiguration>(Configuration);
-            services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
-            services.AddDbContext<ApplicationDbContext>(options =>
-                options.UseSqlServer(
-                    Configuration.GetConnectionString("DefaultConnection")));
-            services.AddDefaultIdentity<IdentityUser>()
-                .AddEntityFrameworkStores<ApplicationDbContext>();
-            services
-                .AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1)
-                .AddJsonOptions(options => options.SerializerSettings.ContractResolver = new DefaultContractResolver()); //prevent JsonResult from camelCasing on its own
 
+            services.AddSingleton(Configuration);
+
+            // services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
+
+            services.AddDbContext<ApplicationDbContext>(options =>
+                    options.UseSqlServer(
+                        Configuration.GetConnectionString("DefaultConnection")))
+                .AddDefaultIdentity<IdentityUser>()
+                .AddDefaultUI()
+                .AddEntityFrameworkStores<ApplicationDbContext>();
+
+
+            services.Configure<FormOptions>(x =>
+            {
+                x.ValueLengthLimit = int.MaxValue;
+                x.MultipartBodyLengthLimit = int.MaxValue;
+                x.MultipartHeadersLengthLimit = int.MaxValue;
+            });
+
+            services
+                .AddMvc()
+                .AddRazorRuntimeCompilation()
+                .AddNewtonsoftJson(options =>
+                {
+                    options.SerializerSettings.ContractResolver = new DefaultContractResolver();
+                    // options.SerializerSettings.Converters.Add(new StringEnumConverter());
+                });
+
+            services.Configure<IdentityOptions>(options =>
+            {
+                // Default Lockout settings.
+                options.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(5);
+                options.Lockout.MaxFailedAccessAttempts = 50;
+                options.Lockout.AllowedForNewUsers = true;
+                options.Password.RequiredLength = 5;
+                options.Password.RequireDigit = false;
+                options.Password.RequireLowercase = false;
+                options.Password.RequireUppercase = false;
+                options.Password.RequiredUniqueChars = 0;
+                options.Password.RequireNonAlphanumeric = false;
+            });
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
-                app.UseDatabaseErrorPage();
-                
             }
             else
             {
                 app.UseExceptionHandler("/Error");
                 app.UseHsts();
             }
+
             app.UseHttpsRedirection();
+            
             app.UseAuthentication();
             // the purpose of the HtmlHandler is to make sure that when people are running SCORM courses (i.e. native html files) that they are authenticated
             app.UseHtmlHandler();
-           
-            app.UseStaticFiles(); // For the wwwroot folder
-            
-            app.UseCookiePolicy();
 
-            app.UseAuthentication();
-           
-            app.UseMvc();
+            app.UseStaticFiles(); // For the wwwroot folder
+
+            app.UseCookiePolicy();
+            
+            app.UseRouting();
+
+            app.UseAuthorization();
+            
+
+            
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapRazorPages();
+                endpoints.MapControllers();
+                // endpoints.MapControllerRoute(
+                //     name: "MyArea",
+                //     pattern: "{area:exists}/{controller=Home}/{action=Index}/{id?}");
+            });
+
             var httpContextAccessor = app.ApplicationServices.GetRequiredService<IHttpContextAccessor>();
+
             Helpers.ConfigurationHelper.Initialize(Configuration, httpContextAccessor);
             // we run SCORM courses out of its own folder (NOT wwwroot) so that HtmlHandler can check for authentication before returning any SCORM content
             app.UseStaticFiles(new StaticFileOptions
             {
                 FileProvider = new PhysicalFileProvider(
-                    Path.Combine(env.ContentRootPath,  Helpers.ConfigurationHelper.CourseFolder)),
+                    Path.Combine(env.ContentRootPath, Helpers.ConfigurationHelper.CourseFolder)),
                 RequestPath = "/SCORM"
             });
-
         }
     }
 }
