@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 
 namespace OpenSourceSCORMLMS.Data
@@ -6,6 +7,6 @@ namespace OpenSourceSCORMLMS.Data
     // this extends the auto-generated ApplicationDbContext with a custom Class for a Query Type
     public partial class ApplicationDbContext : IdentityDbContext
     {
-        public DbQuery<ModelSCORM.SCORM_Course_fromSP> SCORM_Course_FromSP { get; set; }
+        public DbSet<ModelSCORM.SCORM_Course_fromSP> SCORM_Course_FromSP { get; set; }
     }
 }
