@@ -10,17 +10,17 @@
 <h1>Uploaded Packages</h1>
 <table id="gridPackages" class="table table-striped">
     <thead>
-        <tr>
-            <th>Title</th>
-            <th>ModuleID</th>
-            <th>Date Uploaded </th>
-           <th>In study area?</th>  
-            <th>Details</th>
-        </tr>
+    <tr>
+        <th>Title</th>
+        <th>ModuleID</th>
+        <th>Date Uploaded </th>
+        <th>In study area?</th>
+        <th>Details</th>
+    </tr>
     </thead>
     <tbody>
-        @foreach (var item in Model.listPackagesWithUser)
-        {
+    @foreach (var item in Model.listPackagesWithUser)
+    {
         <tr>
             <td>@item.title
 
@@ -30,21 +30,18 @@
             <td>
                 @item.DateUploaded
             </td>
-           
+
             <td>
-                @if (item.user_module_id > 0){
-              <p>Yes</p> 
+                @if (item.user_module_id > 0)
+                {
+                    <p>Yes</p>
                 }
             </td>
-               
+
             <td>
                 <a href="Package?id=@item.id">Package Details</a>
             </td>
         </tr>
-        }
+    }
     </tbody>
 </table>
-
-
-
-
