@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
 namespace OpenSourceSCORMLMS
@@ -9,16 +10,12 @@ namespace OpenSourceSCORMLMS
     {
         public static void Main(string[] args)
         {
-            CreateWebHostBuilder(args).Build().Run();
+            CreateHostBuilder(args).Build().Run();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
+        public static IWebHostBuilder CreateHostBuilder(string[] args) =>
             WebHost.CreateDefaultBuilder(args)
-             .ConfigureLogging((hostingContext, builder) =>
-             {
-                 builder.AddConfiguration(hostingContext.Configuration.GetSection("Logging"));
-                 builder.AddFile(o => o.RootPath = System.AppContext.BaseDirectory);
-             })
+                .ConfigureKestrel((context, options) => { options.Limits.MaxRequestBodySize = 737280000; })
                 .UseStartup<Startup>();
     }
 }
